requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.7.4
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
identify==2.6.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
sentry-sdk==2.12.0
sniffio==1.3.1
SQLAlchemy
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.18.3
virtualenv==20.26.3
